sdr-agent/requirements.txt
googlemaps>=4.10.0
gspread>=5.12.0
google-auth>=2.23.0
